Add the PoC initial version of the screen mirror demo
Add the initial PoC which is working on both i.MX RT1064 and RT1170.
Other side is a Windows / Linux PC using gstreamer to send screen
capture over a TCP/IP network using ENET1G.

This version is initially placed in samples/subsys/video of the Zephyr
code base then moved here as a standalone application. It need to be
adapted to be compiled here.

Network stack code is already relocated on ITCM to optimize performance.

Known issues :
- ENET1G speed is very low (tested with zperf)
- Frame rate is about 3 fps, even worse than expected with the current
ethernet speed

Signed-off-by: Phi Bang Nguyen <phibang.nguyen@nxp.com>