android: Add RemoteDirectConnection
Add a RemoteDirectConnection which connects directly to the remote server
via TCP without passing by an adb tunnel (reverse/forward). It is also
responsible to send video data directly to the remote server through a
TCP socket. Normally, we should separate these two functionalities but
for the sake of simplicity, we do this way for now.

Currently, this works only for JpegCompression, for MediaCodec, some
more modifications will be needed.

This RemoteDirectConnection replaces the DesktopConnection.

Signed-off-by: Phi Bang Nguyen <phibang.nguyen@nxp.com>